commiting all major and minor changes before ripping out the old disassembler and using the new one

+++ Cheat Engine/DBKKernel/interruptHook.h
@@ -46,7 +46,7 @@ typedef
 struct
 {    
 	UINT64 eip;
-	UINT64 cs;
+	WORD cs;
 } JUMPBACK, *PJUMPBACK;
 #pragma pack()
 #else
